requirements.txt
telethon>=1.34.0          # Telegram
discord.py>=2.3.2         # Discord
slackclient>=2.9.4        # Slack RTM
python-dotenv>=1.0.1      # read .env into env vars